Support cached formula string values
Modify ExcelOpenXmlReader to treat "str" cell type the same as "inlineString".
Modify ExcelOpenXmlReader to use the original cell value converted to a string rather than the value parsed as a double.
Add unit test.
Modify the "Issue_11573_BlankValues" test to check for a string instead of a number since string formula values are no longer parsed as numbers.